+++ .framework/node/backend/lib/event.js
@@ -12,13 +12,8 @@ const axios = axiosLib.create({
 });
 
 async function sendEvent(event, metadata) {
-  try {
-    // const result = await axios.get(`/hello`, JSON.stringify({event, metadata}));
-    const result = await axios.post(`/users/${WILCO_ID}/event`, JSON.stringify({event, metadata}));
+    const result = await axios.post(`/users/${WILCO_ID}/event`, JSON.stringify({ event, metadata }));
     return result?.data;
-  } catch (e) {
-    // pass
-  }
 }
 
 module.exports = {
